+++ test/fixtures/basic/app/pages/query-whoami.vue
@@ -1,9 +1,7 @@
 <script setup lang="ts">
-import { useQuery } from '@tanstack/vue-query'
-import { onServerPrefetch } from 'vue'
+import { useQuery } from '@pinia/colada'
 
 const query = useQuery($endpoint('/api/whoami', { method: 'get' }).queryOptions())
-onServerPrefetch(() => query.suspense())
 </script>
 
 <template>